SK:
-ADD: directory Dashboard, which contains all scripts related to
dashboard
-ADD: a common dashboard cmake script, which can be included by each
specific dashboard script. The common script is based on
vxl_common.cmake, which is used by the vxl library in the same way. Vxl
also uses svn, so this script seemed appropriate for elastix. It
implements a few handy things, like initial checkout, sending the ctest
script along with the test results etc. Including this common dashboard
script makes it easier to setup a new dashboard test build.
-ADD: a dashboard script for bigrcluster.erasmusmc, which makes use of
this common dashboard script. It is now running and seems to work.



git-svn-id: https://svn.bigr.nl/elastix/trunkpublic@3452 4d348492-3962-11de-aa85-65a2020054a3